Move IoOwn/IoPtr and related classes out of io-context.h
Continue the effort of decomposing io-context.h/c++ to make it smaller
and more focused.
